+++ src/main/java/refinedstorage/block/BlockCable.java
@@ -237,7 +237,7 @@ public class BlockCable extends BlockCoverable {
         IBlockState state = super.onBlockPlaced(world, pos, facing, hitX, hitY, hitZ, meta, entity);
 
         if (getPlacementType() != null) {
-            return state.withProperty(DIRECTION, getPlacementType().getFrom(facing, entity));
+            return state.withProperty(DIRECTION, getPlacementType().getFrom(facing, pos, entity));
         }
 
         return state;